Organize widget files in folders
