+++ src/motile_plugin/__init__.py
@@ -1,11 +1,7 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from .widgets.motile_widget import MotileWidget
-
 try:
-    __version__ = version("motile-toolbox")
+    __version__ = version("motile-plugin")
 except PackageNotFoundError:
     # package is not installed
     __version__ = "uninstalled"
-
-__all__ = ("MotileWidget",)
