+++ src/motile_plugin/widgets/__init__.py
@@ -1,3 +1,3 @@
-from .motile_widget import MotileWidget  # noqa
-from .tree_widget import TreeWidget  # noqa
-from .tracks_viewer import TracksViewer  # noqa
+from .motile.motile_widget import MotileWidget  # noqa
+from .tracks_view.tree_widget import TreeWidget  # noqa
+from .tracks_view.tracks_viewer import TracksViewer  # noqa
