+++ src/motile_plugin/widgets/motile/motile_widget.py
@@ -16,9 +16,7 @@ from superqt.utils import thread_worker
 from motile_plugin.backend.motile_run import MotileRun
 from motile_plugin.backend.solve import solve
 from motile_plugin.core import Tracks
-from motile_plugin.widgets.tracks_viewer import (
-    TracksViewer,
-)
+from motile_plugin.widgets.tracks_view.tracks_viewer import TracksViewer
 
 from .run_editor import RunEditor
 from .run_viewer import RunViewer
