+++ src/motile_plugin/widgets/tracks_view/tree_view_mode_widget.py
